Add Event Awards link to OBF menu
I was surprised to see that the OBF main menu doesn't include Event Awards - shouldn't it? (Didn't it, at some point?) So I added it, but I don't fully understand how the weights are used, and particularly how menu items are ordered if they lack weights (as some of them do). So this PR may need some tweaking.